+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Ask for user confirmation before downloading large files.
+- Add reprojection functionality to `get_elev_raster` function and `Raster` class.
 
 ## 0.2.0 - 2025-01-06
 ### Added